requirements.txt
Django==1.10.2
PyJWT==1.4.0